+++ docs/ChangeLog.md
@@ -51,6 +51,7 @@
 - syntax: support arbitrary suffixes in `bleopt filename_ls_colors` `#D2213` e169e31d
 - util(vbell): support `bleopt vbell_align=panel` (requested by bb010g) `#D2228` fe85e0dd
 - highlight: reflect the top-level positional parameters `#D2246` xxxxxxxx
+- color: adjust default fg values in faces and add `bleopt color_scheme` (requested by mattmc3) `#D2248`
 
 ## Changes